add fix photo1x2x
