+++ src/js/index.js
@@ -4,4 +4,4 @@ import './components/footer-modal';
 import './firebase/authentication';
 import './components/authentication-modal';
 import './components/pagination-bar';
-import './components/my-library';
+import './components/my-library-btn';
