+++ src/partials/header/header-lib.html
@@ -3,7 +3,11 @@
   <div class="container">
     <ul class="hero-list">
       <li class="hero-list__item">
-        <button class="hero-list__btn is-active" type="button" data-btn-watched>
+        <button
+          class="hero-list__btn is-header-lib-active"
+          type="button"
+          data-btn-watched
+        >
           Watched
         </button>
       </li>
