+++ src/sass/components/footer/_footer-modal.scss
@@ -73,13 +73,16 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    transition: color 250ms $cubic;
+    transform: scale(0.8);
+    transition: color 250ms $cubic, transform 250ms $cubic;
+
     background-color: #f5f4fa;
     color: #000000;
 
     &:hover,
     &:focus {
       color: $background-color-btn;
+      transform: scale(1);
     }
   }
   &__svg {
@@ -97,11 +100,15 @@
     border-radius: 0px 0px 4px 4px;
     background-color: #fff;
     transform: scale(0.9);
-    transition: transform 250ms $cubic;
+    transition: transform 250ms $cubic, box-shadow 250ms $cubic;
 
     &:hover,
     &:focus {
       transform: scale(1);
+      box-shadow: 0 2.8px 2.2px rgba(0, 0, 0, 0.034),
+        0 6.7px 5.3px rgba(0, 0, 0, 0.048), 0 12.5px 10px rgba(0, 0, 0, 0.06),
+        0 22.3px 17.9px rgba(0, 0, 0, 0.072),
+        0 41.8px 33.4px rgba(0, 0, 0, 0.086), 0 100px 80px rgba(0, 0, 0, 0.12);
     }
 
     @media screen and(min-width: 768px) {
