+++ src/sass/components/header/_header-lib.scss
@@ -125,10 +125,11 @@
       width: 147px;
     }
 
-    &.is-active {
+    &.is-header-lib-active {
       filter: drop-shadow(0px 8px 43px rgba(255, 107, 1, 0.6));
       background: #ff6b01;
       border-radius: 5px;
+      border: none;
     }
   }
 }
